Revert 129061 - Create a database for NaCl validation caching that is shared between processes.
Reverted due to perf regression, see 
http://chromegw.corp.google.com/i/chromium/builders/Linux%20x64/builds/25780

This change primarily entails creating a SyncChannel between sel_ldr and the
browser.  Queries to the database could be made from any thread inside sel_ldr,
so the query mechanism needs to be thread safe.

This feature is currently disabled by default, and requires an environment
variable to enable.  A few changes need to be made before this features is safe
and can be enabled, such as making sure each installation has a unique,
crypographically secure key.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2515
TEST= Run NaCl w/ NACL_VALIDATION_CACHE=1


Review URL: http://codereview.chromium.org/9796006

TBR=ncbray@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9808113

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@129077 0039d316-1c4b-4281-b951-d872f2087c98

+++ crypto/symmetric_key_win.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2012 The Chromium Authors. All rights reserved.
+// Copyright (c) 2011 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -50,12 +50,11 @@ ALG_ID GetAESAlgIDForKeySize(size_t key_size_in_bits) {
 // TODO(wtc): use this function in hmac_win.cc.
 bool ImportRawKey(HCRYPTPROV provider,
                   ALG_ID alg,
-                  const void* key_data, size_t key_size,
+                  const void* key_data, DWORD key_size,
                   ScopedHCRYPTKEY* key) {
   DCHECK_GT(key_size, 0);
 
-  DWORD actual_size =
-      static_cast<DWORD>(sizeof(PlaintextBlobHeader) + key_size);
+  DWORD actual_size = sizeof(PlaintextBlobHeader) + key_size;
   std::vector<BYTE> tmp_data(actual_size);
   BYTE* actual_key = &tmp_data[0];
   memcpy(actual_key + sizeof(PlaintextBlobHeader), key_data, key_size);
@@ -67,7 +66,7 @@ bool ImportRawKey(HCRYPTPROV provider,
   key_header->hdr.bVersion = CUR_BLOB_VERSION;
   key_header->hdr.aiKeyAlg = alg;
 
-  key_header->cbKeySize = static_cast<DWORD>(key_size);
+  key_header->cbKeySize = key_size;
 
   HCRYPTKEY unsafe_key = NULL;
   DWORD flags = CRYPT_EXPORTABLE;
@@ -185,7 +184,7 @@ bool GenerateHMACKey(size_t key_size_in_bits,
   if (!ok)
     return false;
 
-  DWORD key_size_in_bytes = static_cast<DWORD>(key_size_in_bits / 8);
+  DWORD key_size_in_bytes = key_size_in_bits / 8;
   scoped_array<BYTE> random(new BYTE[key_size_in_bytes]);
   ok = CryptGenRandom(safe_provider, key_size_in_bytes, random.get());
   if (!ok)
@@ -259,7 +258,7 @@ bool ComputePBKDF2Block(HCRYPTHASH hash,
 
   // Iteration U_1: Compute PRF for S.
   ok = CryptHashData(safe_hash, reinterpret_cast<const BYTE*>(salt.data()),
-                     static_cast<DWORD>(salt.size()), 0);
+                     salt.size(), 0);
   if (!ok)
     return false;
